se editó el proceso de sumar inventario

+++ query Cambios Davis.sql
@@ -1,12 +1,13 @@
 -- procreso de actualizar inventario desde factura - 8/04/2021
 DELIMITER $$
-	create procedure SpActualizarInventarioProductosFacturacion(idBuscado varchar(7), cant int(100),estado tinyint(1))
+	create procedure SpActualizarInventarioProductosFacturacion(idBuscado varchar(7), cant double, estado tinyint(1))
 		BEGIN
 			update InventarioProductos
 				set  estadoProductoId = estado, inventarioproductos.inventarioProductoCant = cant
 					where productoId = idBuscado;
         END $$
 DELIMITER ;
+
 
 
 DELIMITER $$
